requirements/base.txt
Django==1.7.1
PyYAML==3.11
defusedxml==0.4.1
django-countries==2.1.2
django-easy-pjax==1.1.0
django-widget-tweaks==1.3
djangorestframework==2.3.14
psycopg2==2.5.4
python3-openid==3.0.5
pytz==2014.4
requests==2.5.1
requests-oauthlib==0.4.2
mysqlclient==1.3.4

pip==6.0.3
setuptools==10.1
